立即购买逻辑更改；若干已知Bug修复

+++ platform-api/src/main/java/com/platform/service/ApiOrderService.java
@@ -211,7 +211,9 @@ public class ApiOrderService {
         }
 
         //清空已购买的商品
-        apiCartMapper.deleteByCart(loginUser.getUserId(), 1, 1);
+        if (type.equals("cart")) {
+            apiCartMapper.deleteByCart(loginUser.getUserId(), 1, 1);
+        }
         resultObj.put("errno", 0);
         resultObj.put("errmsg", "订单提交成功");
         //
